+++ sdk/log/simple_test.go
@@ -6,6 +6,7 @@ package log_test
 import (
 	"context"
 	"io"
+	"strconv"
 	"strings"
 	"sync"
 	"testing"
@@ -13,7 +14,17 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/sdk"
+	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/log"
+	"go.opentelemetry.io/otel/sdk/log/internal/observ"
+	"go.opentelemetry.io/otel/sdk/metric"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata"
+	"go.opentelemetry.io/otel/sdk/metric/metricdata/metricdatatest"
+	semconv "go.opentelemetry.io/otel/semconv/v1.37.0"
+	"go.opentelemetry.io/otel/semconv/v1.37.0/otelconv"
 )
 
 type exporter struct {
@@ -38,6 +49,17 @@ func (e *exporter) Shutdown(context.Context) error {
 func (e *exporter) ForceFlush(context.Context) error {
 	e.forceFlushCalled = true
 	return nil
+}
+
+var _ log.Exporter = (*failingTestExporter)(nil)
+
+type failingTestExporter struct {
+	exporter
+}
+
+func (f *failingTestExporter) Export(ctx context.Context, r []log.Record) error {
+	_ = f.exporter.Export(ctx, r)
+	return assert.AnError
 }
 
 func TestSimpleProcessorOnEmit(t *testing.T) {
@@ -137,4 +159,173 @@ func BenchmarkSimpleProcessorOnEmit(b *testing.B) {
 
 		_ = out
 	})
+}
+
+func BenchmarkSimpleProcessorObservability(b *testing.B) {
+	run := func(b *testing.B) {
+		slp := log.NewSimpleProcessor(&failingTestExporter{exporter: exporter{}})
+		record := new(log.Record)
+		record.SetSeverityText("test")
+
+		ctx := b.Context()
+		b.ReportAllocs()
+		b.ResetTimer()
+
+		var err error
+		for b.Loop() {
+			err = slp.OnEmit(ctx, record)
+		}
+		_ = err
+	}
+
+	b.Run("Observability", func(b *testing.B) {
+		b.Setenv("OTEL_GO_X_OBSERVABILITY", "true")
+		run(b)
+	})
+	b.Run("NoObservability", run)
+}
+
+func TestSimpleLogProcessorObservability(t *testing.T) {
+	testcases := []struct {
+		name          string
+		enabled       bool
+		exporter      log.Exporter
+		wantErr       error
+		assertMetrics func(t *testing.T, rm metricdata.ResourceMetrics)
+	}{
+		{
+			name:     "disabled",
+			enabled:  false,
+			exporter: new(exporter),
+			assertMetrics: func(t *testing.T, rm metricdata.ResourceMetrics) {
+				assert.Empty(t, rm.ScopeMetrics)
+			},
+		},
+		{
+			name:     "enabled",
+			enabled:  true,
+			exporter: new(exporter),
+			assertMetrics: func(t *testing.T, rm metricdata.ResourceMetrics) {
+				assert.Len(t, rm.ScopeMetrics, 1)
+				sm := rm.ScopeMetrics[0]
+
+				p := otelconv.SDKProcessorLogProcessed{}
+
+				want := metricdata.ScopeMetrics{
+					Scope: instrumentation.Scope{
+						Name:      observ.ScopeName,
+						Version:   sdk.Version(),
+						SchemaURL: semconv.SchemaURL,
+					},
+					Metrics: []metricdata.Metrics{
+						{
+							Name:        p.Name(),
+							Description: p.Description(),
+							Unit:        p.Unit(),
+							Data: metricdata.Sum[int64]{
+								DataPoints: []metricdata.DataPoint[int64]{
+									{
+										Value: 1,
+										Attributes: attribute.NewSet(
+											observ.GetSLPComponentName(0),
+											semconv.OTelComponentTypeKey.String(
+												string(otelconv.ComponentTypeSimpleLogProcessor),
+											),
+										),
+									},
+								},
+								Temporality: metricdata.CumulativeTemporality,
+								IsMonotonic: true,
+							},
+						},
+					},
+				}
+
+				metricdatatest.AssertEqual(
+					t,
+					want,
+					sm,
+					metricdatatest.IgnoreExemplars(),
+					metricdatatest.IgnoreTimestamp(),
+				)
+			},
+		},
+		{
+			name:    "Enable Exporter error",
+			enabled: true,
+			wantErr: assert.AnError,
+			exporter: &failingTestExporter{
+				exporter: exporter{},
+			},
+			assertMetrics: func(t *testing.T, rm metricdata.ResourceMetrics) {
+				assert.Len(t, rm.ScopeMetrics, 1)
+				sm := rm.ScopeMetrics[0]
+				p := otelconv.SDKProcessorLogProcessed{}
+
+				want := metricdata.ScopeMetrics{
+					Scope: instrumentation.Scope{
+						Name:      "go.opentelemetry.io/otel/sdk/log/internal/observ",
+						Version:   sdk.Version(),
+						SchemaURL: semconv.SchemaURL,
+					},
+					Metrics: []metricdata.Metrics{
+						{
+							Name:        p.Name(),
+							Description: p.Description(),
+							Unit:        p.Unit(),
+							Data: metricdata.Sum[int64]{
+								DataPoints: []metricdata.DataPoint[int64]{
+									{
+										Value: 1,
+										Attributes: attribute.NewSet(
+											observ.GetSLPComponentName(0),
+											semconv.OTelComponentTypeKey.String(
+												string(otelconv.ComponentTypeSimpleLogProcessor),
+											),
+											semconv.ErrorTypeKey.String("*errors.errorString"),
+										),
+									},
+								},
+								Temporality: metricdata.CumulativeTemporality,
+								IsMonotonic: true,
+							},
+						},
+					},
+				}
+
+				metricdatatest.AssertEqual(
+					t,
+					want,
+					sm,
+					metricdatatest.IgnoreTimestamp(),
+					metricdatatest.IgnoreExemplars(),
+				)
+			},
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			t.Setenv("OTEL_GO_X_OBSERVABILITY", strconv.FormatBool(tc.enabled))
+
+			original := otel.GetMeterProvider()
+			t.Cleanup(func() {
+				otel.SetMeterProvider(original)
+			})
+
+			r := metric.NewManualReader()
+			mp := metric.NewMeterProvider(metric.WithReader(r))
+			otel.SetMeterProvider(mp)
+
+			slp := log.NewSimpleProcessor(tc.exporter)
+			record := new(log.Record)
+			record.SetSeverityText("test")
+			err := slp.OnEmit(t.Context(), record)
+			require.ErrorIs(t, err, tc.wantErr)
+			var rm metricdata.ResourceMetrics
+			require.NoError(t, r.Collect(t.Context(), &rm))
+			tc.assertMetrics(t, rm)
+			observ.SetSimpleProcessorID(0)
+		})
+	}
 }